requirements.txt
amqp==1.4.6
anyjson==0.3.3
appdirs==1.4.0
beautifulsoup4==4.5.3
billiard==3.3.0.20
bs4==0.0.1
celery==3.1.18
Django==1.8.2
django-celery==3.2.1
kombu==3.0.30
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pkg-resources==0.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2015.4
redis==2.10.3
requests==2.7.0
simplejson==3.10.0
six==1.10.0
